Refactors enigma class to work with encrypt runner file
