+++ test/encrypt_test.rb
@@ -1,6 +1,6 @@
 require './test/test_helper'
 
-class EncryptTest < Minitest::Test
+class EncryptingTest < Minitest::Test
   def test_it_exists
     encrypt = Encrypt.new
     assert_instance_of Encrypt, encrypt
