+++ test/test_helper.rb
@@ -4,5 +4,5 @@ require 'minitest/autorun'
 require 'minitest/pride'
 require 'mocha/minitest'
 require './lib/enigma'
-require './lib/encrypt'
+require './lib/encrypting'
 require './lib/shiftable'
